# [Enhancement] Consider key container memory usages for streaming aggregate