## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR relates：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Closes #6212.

## Problem Summary：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->
Streaming Agg decides whether to PreAgg according to the memory usage of Hash Table and the aggregation degree. When the HT memory is less than 256KB, PreAgg will be always used.

For now, HT memory only counts the memory allocated by `MemPool`. However, in some scenarios, in order to optimize performance, HT's keys don't allocate from MemPool, such as `AggHashSetOfSerializedKeyFixedSize`, which will cause `MemPool` to be always 0. Therefore, even if there is no degree of aggregation, PreAgg will always be used.

Therefore, this PR consider the key memory used for streaming aggregate.
- If keys are allocated from `MemPool`, the key memory usage is the pointer of the key.
- Otherwise, the key memory usage is the real key memory usage.


## Test Result
### Environment
- 3BE with 64GB memory and 16vCPUs.

### Query
```sql
-- Q1 Stremaing Distinct Aggregate
select count(distinct lo_orderkey*10+lo_linenumber) from lineorder;

-- Q2 Stremaing Distinct Aggregate
select count(*) from (
    SELECT LO_CUSTKEY, LO_SUPPKEY, sum(LO_REVENUE) 
    FROM lineorder_flat 
    GROUP BY LO_CUSTKEY, LO_SUPPKEY
) AS a;

-- Q3 Stremaing Distinct Aggregate
select count(*) from (
    SELECT LO_PARTKEY, LO_LINENUMBER, LO_DISCOUNT, sum(LO_REVENUE) 
    FROM lineorder_flat 
    GROUP BY LO_PARTKEY, LO_LINENUMBER, LO_DISCOUNT 
) AS a;

-- Q4 Stremaing Aggregate
select sum(LO_REVENUE) 
from lineorder 
group by LO_PARTKEY, LO_LINENUMBER, LO_DISCOUNT 
limit 10;
```

### Result
| Name    | Before (s) | After (s) | After/Before |
| ------- | ---------- | --------- | ------------ |
| query01 | 10.01      | 6.36      | 0.64         |
| query02 | 9.77       | 6.17      | 0.63         |
| query03 | 5.99       | 3.51      | 0.59         |
| query04 | 4.02       | 3.72      | 0.93         |
| SUM     | 29.79      | 19.76     | 0.66         |